Policy Dataloader (#160)
* init vhus

* remove old files

* add vhus vocab

* move usermodule in vhus

* vhus for camrest

* fix informable-keys

* fix vhus user manager

* revise config

* sclstm for multiwoz

* move model dir

* sclstm for camrest

* git ignore

* run sclstm in camrest

* sclstm data for camrest

* sclstm: update camrest model

* sclstm download from tatk

* fix

* pseudo policy training script

* update interface for agenda policy

* update interface for vhus policy

* add env

* ppo

* pg

* config for rl policy

* __init__ for rl policy

* README of training RL policy

* init README of policy module

* update policy README

* load model for mle & vhus

* data split in mle & vhus

* from BERT to SVM in example script

* rename

* update README

* update README

* update README

* test for mle policy

* README for mle policy

* init vhus

* remove old files

* move usermodule in vhus

* vhus for camrest

* fix vhus user manager

* update README

* update README

* update README

* test for mle policy

* README for mle policy

* update README

* update README

* update metrics

* update README

* fix conflict

* fix vectorize

* fix vectorize

* fix MLE policy

* is_train for RL policy

* update README

* rm conflict file

* fix VHUS on camrest

* update result

* update camrest data

* fix sclstm on camrest

* performance by table

* add info about models

* update SCLSTM README

* add evaluator into sess

* update README

* init policy dataloader

* framework: dialog act level policy

* add user dialog policy dataloader & abstract mle policy

* simplify vhus policy
