+++ tatk/policy/mle/__init__.py
@@ -1,0 +1,1 @@
+from tatk.policy.policy import Policy
