+++ tatk/util/dataloader/dataset_dataloader.py
@@ -45,7 +45,9 @@ class MultiWOZDataloader(DatasetDataloader):
                   last_opponent_utterance=False,
                   last_self_utterance=False,
                   session_id=False,
-                  span_info=False
+                  span_info=False,
+                  terminal=False,
+                  goal=False
                   ):
 
         def da2tuples(dialog_act):
@@ -103,6 +105,10 @@ class MultiWOZDataloader(DatasetDataloader):
                         self.data[data_key]['session_id'].append(sess_id)
                     if span_info:
                         self.data[data_key]['span_info'].append(turn['span_info'])
+                    if terminal:
+                        self.data[data_key]['terminal'].append(i+2 >= len(sess['log']))
+                    if goal:
+                        self.data[data_key]['key'].append(sess['goal'])
                     cur_context.append(text)
                     cur_context_dialog_act.append(da)
         return self.data
