+++ tatk/util/dataloader/module_dataloader.py
@@ -27,6 +27,31 @@ class MultiTurnNLUDataloader(ModuleDataloader):
         kwargs.setdefault('context_window_size', 3)
         return self.dataset_dataloader.load_data(*args, **kwargs)
 
+class ActPolicyDataloader(ModuleDataloader):
+    def load_data(self, *args, **kwargs):
+        kwargs.setdefault('belief_state', True)
+        kwargs.setdefault('dialog_act', True)
+        kwargs.setdefault('terminal', True)
+        kwargs.setdefault('context_dialog_act', True)
+        kwargs.setdefault('context_window_size', 2)
+        return self.dataset_dataloader.load_data(*args, **kwargs)
+    
+class ActUserPolicyDataloader(ModuleDataloader):
+    def load_data(self, *args, **kwargs):
+        kwargs.setdefault('goal', True)
+        kwargs.setdefault('dialog_act', True)
+        kwargs.setdefault('terminal', True)
+        kwargs.setdefault('context_dialog_act', True)
+        kwargs.setdefault('context_window_size', 2)
+        return self.dataset_dataloader.load_data(*args, **kwargs)
+
+class WordPolicyDataloader(ModuleDataloader):
+    def load_data(self, *args, **kwargs):
+        kwargs.setdefault('belief_state', True)
+        kwargs.setdefault('utterance', True)
+        kwargs.setdefault('context', True)
+        kwargs.setdefault('context_window_size', 3)
+        return self.dataset_dataloader.load_data(*args, **kwargs)
 
 class SingleTurnNLGDataloader(ModuleDataloader):
     def load_data(self, *args, **kwargs):
